requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.0
# Optional: better token estimation
# tiktoken==0.6.0